// dashboard/src/generated/trade-bot-dashboard/book-level.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from 'flatbuffers';

export class BookLevel {
  bb: flatbuffers.ByteBuffer|null = null;
  bb_pos = 0;
  __init(i:number, bb:flatbuffers.ByteBuffer):BookLevel {
  this.bb_pos = i;
  this.bb = bb;
  return this;
}

price():number {
  return this.bb!.readFloat64(this.bb_pos);
}

size():number {
  return this.bb!.readFloat64(this.bb_pos + 8);
}

static sizeOf():number {
  return 16;
}

static createBookLevel(builder:flatbuffers.Builder, price: number, size: number):flatbuffers.Offset {
  builder.prep(8, 16);
  builder.writeFloat64(size);
  builder.writeFloat64(price);
  return builder.offset();
}

}