// dashboard/src/generated/trade-bot-dashboard/equity-point.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from 'flatbuffers';

export class EquityPoint {
  bb: flatbuffers.ByteBuffer|null = null;
  bb_pos = 0;
  __init(i:number, bb:flatbuffers.ByteBuffer):EquityPoint {
  this.bb_pos = i;
  this.bb = bb;
  return this;
}

ts():bigint {
  return this.bb!.readInt64(this.bb_pos);
}

equity():number {
  return this.bb!.readFloat64(this.bb_pos + 8);
}

static sizeOf():number {
  return 16;
}

static createEquityPoint(builder:flatbuffers.Builder, ts: bigint, equity: number):flatbuffers.Offset {
  builder.prep(8, 16);
  builder.writeFloat64(equity);
  builder.writeInt64(BigInt(ts ?? 0));
  return builder.offset();
}

}